Revert "Remove canadian stage one"
This reverts commit 097e41e7aa424f2f26ed02d8f1829caf80cd7669.

+++ makes/src/40-expat.sh
@@ -3,6 +3,10 @@
 # shellcheck source=./common.sh
 source "$(dirname "$0")/common.sh"
 
+if [ "$CANADIAN_STAGE_ONE" = "true" ]; then
+    exit 0
+fi
+
 rm -rf "${BUILD_DIR}/expat-build"
 mkdir "${BUILD_DIR}/expat-build"
